Introduce backlog processing
If a multi-member cluster is bootstrapped at the same time, one node is
bootstrapping the cluster, while the other nodes wait to be approved. In that
case it is not sufficient to watch for changing keys, since the watch process
is active after the join request has been issied. Instead we look for a backlog
of nodes.